+++ src/kinetis.c
@@ -54,9 +54,9 @@
 
 #define KL25_PAGESIZE 0x400
 
-static int kl25_flash_erase(struct target_s *target, uint32_t addr, int len);
+static int kl25_flash_erase(struct target_s *target, uint32_t addr, size_t len);
 static int kl25_flash_write(struct target_s *target, uint32_t dest,
-			  const uint8_t *src, int len);
+                            const uint8_t *src, size_t len);
 
 static const char kl25_xml_memory_map[] = "<?xml version=\"1.0\"?>"
 /*	"<!DOCTYPE memory-map "
@@ -121,7 +121,7 @@ kl25_command(struct target_s *t, uint8_t cmd, uint32_t addr, const uint8_t data[
 	return true;
 }
 
-static int kl25_flash_erase(struct target_s *t, uint32_t addr, int len)
+static int kl25_flash_erase(struct target_s *t, uint32_t addr, size_t len)
 {
 	addr &= ~(KL25_PAGESIZE - 1);
 	len = (len + KL25_PAGESIZE - 1) & ~(KL25_PAGESIZE - 1);
@@ -135,7 +135,7 @@ static int kl25_flash_erase(struct target_s *t, uint32_t addr, int len)
 }
 
 static int kl25_flash_write(struct target_s *t, uint32_t dest,
-			  const uint8_t *src, int len)
+			  const uint8_t *src, size_t len)
 {
 	/* FIXME handle misaligned start and end of sections */
 	if ((dest & 3) || (len & 3))
